Better workspace/project and scheme/target selector
"Better" in terms that it allows working both with targets and schemes,
but actually it is quite a dirty way to do it.

It should be done by a separate workspace and project construction as,
for example, the workspace doesn't allow target specification, only
schemes.

Also it is not quite Ruby-native :-) As I don't know Ruby good enough to
make it correctly in Ruby-style